getTotalVotes function

+++ packages/revolution-contracts/src/governance/VerbsDAOInterfaces.sol
@@ -228,7 +228,7 @@ contract VerbsDAOStorageV1 is VerbsDAOProxyStorage {
         /// @notice Whether or not the voter supports the proposal or abstains
         uint8 support;
         /// @notice The number of votes the voter had, which were cast
-        uint96 votes;
+        uint256 votes;
     }
 
     /// @notice Possible states that a proposal may be in
